Quad module written
The quad module performs Clenshaw-Curtis quadrature for an input function as part of calculating the Gram matrices. Precompile calls have been added to the beginning of the module files for faster loading in the main program. Data I/O moved to main program as mutable data structure for Inputs was causing performance issues